refactor(launcher): lift startup, install and chrome helpers out of index.ts
index.ts was 3494 lines of everything the main process does. The six
concerns below had no reason to live next to the IPC layer, so they now
have files of their own and index.ts is down to 2689 lines:

  bootstrap/startup-log   startup log rotation, fatal-error reporting
  bootstrap/node-runtime  Node/npm download, checksum, extract, PATH fix
  install-progress        install phase inference + user-facing errors
  window-chrome           theme source, titlebar overlay, tray icon, splash
  net-config              download region, npm registry probe, proxy
  web-security            external-link policy, window navigation guards

Three things were deliberately left where they were:

- The `process.on("uncaughtException"/"unhandledRejection")` registration
  stays at module scope in index.ts. Its whole point is covering the gap
  between `require` and `whenReady`; moving it into the module would have
  changed when it takes effect. Only the handler function moved.
- The pre-ready sequence (setName → appendSwitch →
  disableHardwareAcceleration → globalPaths.push) is untouched, and none
  of the new modules run anything at import time.
- InstallProgress takes `() => mainWindow` rather than the window itself.
  mainWindow is reassigned when the window is closed and reopened from
  the tray, so a captured reference would leave install progress silently
  broken after the first reopen.

Also removes dead imports: crypto/pipeline/Transform were unused before
this change, and `dialog` only looked unused because three call sites
shadowed it with a local require("electron") — those requires are gone
instead, so the top-level import is what's used.

Verified as a pure move: the 124 IPC channels and AgentManager's 79
public methods are identical to before, every extracted function was
diffed against its original, the bundle still contains all of it, and a
throwaway smoke test exercised the wiring at runtime. tsc and the test
suite match the pre-refactor baseline (the 2 download.test.ts network
timeouts were already failing).
